plugin force working
issue was that px4 is trying to stabilize the drone all the time, so if we try to input an external force using LinearVelCmd, it won't work. Workaround is apparently to use ExternalWorldWrench. Moreover, this topic is in world frame ENU (GPT source, so to check), whereas LinVelCmd was in local vehicle frame